Refactor to support ADR004 module exporting
